Conversation and Funnels api integrations completed. Pending Mockito tests
FunnelSensai Endpoints mapping GoHighLevel Api  for the corresponding integrations( Fetch list of Conversations and [Fetch List of Funnels, Fetch List of funnel pages, Fetch count of funnel pages]) is completed. That means Dto, service class and controller. For Fetch list of Conversation is being included a repo and an entity for Db save as testing purpose. Pending include Mockito tests